Support creating a new workspace client from an account client (#792)
## Changes
This PR introduces `func (a *AccountClient) GetWorkspaceClient(w
provisioning.Workspace) (*WorkspaceClient, error)` to make it easy to
get a workspace client for a given workspace. The workspace client uses
a copy of the config from the account client, reusing the underlying
authenticator as well. This allows all workspace clients to use the same
underlying token when authenticating with OAuth, decreasing the number
of unnecessary token refreshes. Additionally, it allows workspace
clients to work with U2M OAuth using the account-level token for
customers with Unified Login.

Use-case:
https://github.com/databricks/terraform-provider-databricks/pull/3188

## Tests
New example passes. 

- [ ] `make test` passing
- [ ] `make fmt` applied
- [ ] relevant integration tests applied