Add `CurrentWorkspaceID()` method on WorkspaceClient (#808)
## Changes
There are times when it is especially useful to get the workspace ID for
the current workspace client. Currently, the workspace ID for the
current workspace is exposed as a header in the SCIM Me API call. We'll
expose this through a `GetWorkspaceId()` method, caching the workspace
ID for the lifetime of the client.

In the future, we may add a `meta` service for exposing information
about the current account/workspace. At that point, we can migrate off
of this somewhat hacky approach.

## Tests
<!-- 
How is this tested? Please see the checklist below and also describe any
other relevant tests
-->

- [ ] `make test` passing
- [ ] `make fmt` applied
- [ ] relevant integration tests applied