Fix GetWorkspaceClient for GCP (#803)
## Changes
The current implementation of GetWorkspaceClient copies the entire
config, critically reusing the cached authenticator so as to use the
same auth mechanism when getting a workspace client. However, at least
in GCP, account-level OAuth tokens can't be used to authenticate to a
workspace (probably because the audience for the account-level and
workspace-level tokens is different).

This PR fixes this by only copying the exported fields and not copying
the authenticator or refresh client. Subsequent use of the config in
WorkspaceClient will trigger config resolution. For GCP, this means
creating a new token source using the correct host as the audience.

## Tests
Manually ran this integration test in all non-UC (Azure, AWS, GCP) and
UC (AWS, GCP) account-level environments.

- [ ] `make test` passing
- [ ] `make fmt` applied
- [ ] relevant integration tests applied