Differentiate between empty structures and components that can take on any value (#821)
## Changes
Currently, the SDK doesn't distinguish between empty structures and
components that can take any value. As a result, the SDK uses type `any`
for fields that should really have a type of `struct{}` or `*struct{}`
(or an equivalent type). This PR updates the OpenAPI generator code to
distinguish between these two cases.

Additionally, we update the generation of SDK structures to include
empty types. This includes some types that are not currently used in the
SDK, mainly empty responses because the generated API calls for methods
with an empty response just omits the response return value entirely.

Two changes we should consider for later:
1. Include empty response types in the return value for API calls. Even
though they are empty, it means that adding anything to an empty
response is a breaking change for SDK users.
2. Combine `emptyTypes` and `types`. Users can always filter out empty
types based on `Entity.IsEmpty`. This reduces unnecessary complexity in
the generator. However, first, the other SDKs need to be updated to
handle this case.

This also includes some changes from the Settings team and Files team,
but it is necessary to update as there were issues with structures that
were incorrectly annotated with `x-databricks-any`.

## Tests
<!-- 
How is this tested? Please see the checklist below and also describe any
other relevant tests
-->

- [ ] `make test` passing
- [ ] `make fmt` applied
- [ ] relevant integration tests applied