Treat empty entities as entities (#831)
## Changes
To simplify codegen for Java, we'll start treating empty types as normal
entities in the OpenAPI code generator. The `(e *Entity) IsEmpty()` can
be used to determine if an entity is empty or not.

## Tests
<!-- 
How is this tested? Please see the checklist below and also describe any
other relevant tests
-->

- [ ] `make test` passing
- [ ] `make fmt` applied
- [ ] relevant integration tests applied