Adaptive request timeouts
The net/http package supports only a fixed timeout.

This is problematic when streaming a large request or response body.

The approach proposed in this PR is to disable the net/http timeouts and use
our own context for cancellation instead. This context is cancelled only if
there is no activity for the duration of 1 timeout period. As the request or
response body is read, the deadline is reinitialized to equal 1 timeout period
from the current timestamp. This means real timeouts will still timeout in the
same amount of time, but streaming requests will not timeout as long as there
is activity.