Adaptive request timeouts (#837)
## Changes

The net/http package supports only a fixed timeout.

This is problematic when streaming a large request or response body.

The approach proposed in this PR is to disable the net/http timeouts and
use our own context for cancellation instead. This context is canceled
only if there is no activity for the duration of 1 timeout period. As
the request or response body is read, the deadline is reinitialized to
equal 1 timeout period from the current timestamp. This means real
timeouts will still time out in the same amount of time, but streaming
requests will not time out as long as there is activity.

If a request times out, it used to return the following error:
> Get "/a?": context deadline exceeded (Client.Timeout exceeded while
awaiting headers)

Now it returns the following error:
> Get "/a?": request timed out after 10ms of inactivity

**Note:** the unit tests are time-based. Not great but I don't have a
great alternative (that doesn't involve mocking time itself).

## Tests

- [x] `make test` passing
- [x] `make fmt` applied
- [x] relevant integration tests applied

---------

Co-authored-by: Miles Yucht <miles@databricks.com>