Add retries for "worker env WorkerEnvId(workerenv-XXXXX) not found" (#890)
## Changes
Due to the eventually consistent nature of worker environment creation
as a part of workspace setup, certain API requests can fail when made
right after workspace creation. We have some exception handling for
these in the SDKs, but a new case has appeared: "worker env
WorkerEnvId(workerenv-XXXXX) not found" (see
https://github.com/databricks/terraform-provider-databricks/issues/3452).

This PR addresses this issue. Furthermore, it moves the transient error
messages into autogeneration so that there is a single source of truth
that applies to all SDKs.

One other small change: I removed running unit tests from the
autogeneration flow. It removes a small amount of convenience (users
need to run `make test` after regenerating the SDK), but it speeds up
the devloop when iterating on code generation, and it allows the release
flow to continue to make a PR before failing. This PR can be modified as
needed to fix any test failures or compilation failures, as per usual.

## Tests
Added a unit test to cover this.

- [ ] `make test` passing
- [ ] `make fmt` applied
- [ ] relevant integration tests applied