Allow unlimited timeouts in retries (#904)
## Changes
Terraform controls the timeout of resource operations through methods
like `CreateContext`. The default retry timeout at the client level is 5
minutes, and it applies to all operations. Terraform has a default
timeout of 20m, but users can override the timeout for individual
resources in TF templates
(https://developer.hashicorp.com/terraform/plugin/sdkv2/resources/retries-and-customizable-timeouts).
Terraform ensures that timeouts are enforced by creating a context with
timeout before calling the provider's implementation for a given
resource or datasource. To ensure that the provider can use the full
timeout period before cancelling, this change allows a user to configure
infinite timeouts for retries by setting `RETRY_TIMEOUT_SECONDS` to a
negative value.

## Tests
It's hard to test this via a unit test, but I ran the TF provider with
this Go SDK and manually verified that the Go SDK didn't cancel requests
within 5 minutes but rather used the entire timeout at the resource
level.

- [ ] `make test` passing
- [ ] `make fmt` applied
- [ ] relevant integration tests applied