Log cancelled and failed requests (#919)
## Changes
Currently, we do not log requests that failed to complete or that failed
when the SDK started processing the response. This means that requests
that timeout due to a client-side timeout or canceled context are not
logged, meaning that there is no visibility in the log of these kinds of
requests.

This PR changes the request processing and error handling to ensure that
all requests are logged, even those that are cancelled or fail due to
other errors. The general approach is to process as much of the response
as possible until encountering an error, log what we've retrieved, and
then do any post-processing.

Some changes are needed in the roundtrip stringer to handle the case of
a nil response.

## Tests
Unit tests verify that canceled requests or nil responses are logged.

- [ ] `make test` passing
- [ ] `make fmt` applied
- [ ] relevant integration tests applied