Support partners in SDK (#925)
## Changes
Partners of Databricks need a mechanism to register themselves in
libraries or applications that they write. In this way, requests made by
users of those libraries will include sufficient information to link
those requests to the original users.

This PR adds a new function `WithPartner` that can be used by a partner
to add partner information to the User-Agent header for requests made by
the SDK. The new header will have the form `partner/<parther id>`. The
partner identifier is opaque for the SDK, but it must be alphanumeric.

This PR also lifts the requirement that there may only be a single copy
of an entry in the user agent. This allows multiple partners to register
themselves in the same library or application.

## Tests
Unit tests for user agent were modified to support multiple entries for
the same header key. A unit test is added to cover the WithProduct
method.

- [ ] `make test` passing
- [ ] `make fmt` applied
- [ ] relevant integration tests applied