Parse API Error messages with `int` error codes (#960)
## Changes

Parsing error messages from the `/api/2.0/preview/scim/v2/Me` API causes
parsing errors because the API returns the HTTP status (i.e. an `int`)
as error code instead of a `string`. For example:

```json
{
   "error_code": 403,
   "message": "Invalid access token."
}
```

This PR solves the problem by parsing the `error_code` field as a
`string` no matter its type.

Note that the implementation is overkill because it also converts _any_
literal to a string (e.g. `true` becomes `"true"`). We could guarantee
that the behavior is limited to converting `int` to `string` by adding a
custom [Unmarshaler](https://pkg.go.dev/encoding/json#Unmarshaler) but
this didn't seem worth the additional code.

This PR also:

- Refactors the code by removing the `APIErroBody` struct. The rationale
is that the struct exists solely for parsing and is actually not used
outside of the `parseErrorFromResponse` function. Turning it into an
anonymous struct within `parseErrorFromResponse` makes the intent
crystal clear.
- Slightly restructures the test cases to better match go's naming
conventions and separate the "three tests parts" (init, test, assert).

## Tests

Added a unit test to verify that (i) an int error code are parsed as
string, and (ii) that this string is properly ignored.

- [x] `make test` passing
- [x] `make fmt` applied
- [x] relevant integration tests applied