Add a credentials provider for Github Azure OIDC (#965)
_Note: this PR is a copy of PR #950 which could not be merged because of
some unverified commits. Please check PR #950 for the original review
and comments._

## Changes

This PR adds a `CredentialsProvider` to authenticate with Azure from
Github workflows.

The code is inspired by a similar feature already implemented in the
Python SDK. It works as follows:

1. Obtain an ID token from Azure leveraging the env variables
`ACTIONS_ID_TOKEN_REQUEST_URL` and `ACTIONS_ID_TOKEN_REQUEST_TOKEN` as
[explained
here](https://docs.github.com/en/actions/deployment/security-hardening-your-deployments/configuring-openid-connect-in-cloud-providers).
2. Exchange that ID token for an auth token.

## Tests

Added a test suite which covers all the added code paths. I've also
confirmed in my own Github Action that the code is properly able to
authenticate.

Note: I'm not super happy with how errors are compared (i.e. using a
prefix) which is a little brittle. A better approach would be to
leverage `errors.As` or `errors.Is`. However, it is difficult to do that
at the moment without adding ad hoc new error types. A longer term
solution would probably involve standardizing the package around a set
of clearly defined error types shared by all implementations of
`CredentialsProvider` in `config`. That is out of the scope of this PR
though.

- [x] `make test` passing
- [x] `make fmt` applied
- [x] relevant integration tests applied