[Internal] Log a warning when declaring inline entities  (#994)
## Changes

Inlined anonymous entities (object and enum) is discouraged. Rather, the
proper way to define these entities is to declare them explicitly in the
component part of the OpenAPI specification.

This PR adds a warning to inform users that their schema contains the
discouraged pattern.

This PR also includes a couple of noop refactoring to improve
readability:
- Move the part responsible for handling references in its own function.
- Use a switch to emphasize that an entity can only be one of a fixed
set of types.

- [x] `make test` passing
- [x] `make fmt` applied
- [x] relevant integration tests applied