[Internal] Prefix all extension files with `ext_` (#1001)
## Changes

This PR clearly differentiate auto-generated files from extension files
by prefixing all extension files with `ext_`. An extension file is a
service package file that hasn't been auto-generated and provides
additional functionalities to the services in its package.

An alternative solution would be to prefix all auto-generated files with
`gen_`. The reason I decided to go for `ext_` are (i) it requires less
work, and (ii) it doesn't prevent us from moving to the `gen_` solution
in the future if we want to.

This PR also removes a `compute/compliance.go` which only contained
commented code.

## Tests

Normal unit and integration tests to assert that no functionality got
removed by accident.

- [x] `make test` passing
- [x] `make fmt` applied
- [x] relevant integration tests applied