[Feature] Remove deprecated `WithImpl` and `Impl` service methods (#1003)
## Changes

This PR removes the deprecated `WithImpl` and `Impl` method so that the
`ServiceAPI` struct can embed the `serviceImpl` auto-generated
implementation directly. This has two advantages. First, it guarantees
that proper go doc can be generated by showing the promoted methods on
the embedding struct. Second, it simplifies some extensions by removing
the need to cast the interface to the implementation struct.

**Important**: this is a breaking change. Although, it is considered OK
in this case as (i) the SDKs are still in an experimental phase, and
(ii) the methods have been flagged as deprecated for a long time.

## Tests

- [x] `make test` passing
- [x] `make fmt` applied
- [x] relevant integration tests applied