[Internal] Fix integration tests (#1008)
## Changes
There are three issues affecting our integration tests today:
1. When authenticated as a service principal, GetWorkspaceClient only
works if the target workspace is a UC-enabled workspace. Currently, the
test targets a non-UC workspace, so I've changed it to target a UC
workspace.
2. In our AWS environment, clusters/warehouses can only be created if
they have an "Owner" tag. I've updated TestAccSqlWarehouses to add this
tag corresponding to our team.
3. There was a change in the behavior of cluster listing. Previously,
list clusters API would return recently terminated clusters. However,
now that doesn't seem to be the case. I've raised this with the
appropriate backend team. In the meantime, I've updated the integration
test to do all listing operations while the cluster is running.

## Tests
Triggered the integration tests on this PR.

- [ ] `make test` passing
- [ ] `make fmt` applied
- [ ] relevant integration tests applied