[Fix] Handle non-JSON errors (#1031)
## Changes
Some errors returned by the platform are not serialized using JSON (see
https://github.com/databricks/databricks-sdk-go/issues/998 for an
example). They are instead serialized in the form "<ERROR_CODE>:
<MESSAGE>". Today, the SDK cannot parse these error messages well,
resulting in a poor user experience.

This PR adds support for parsing these error messages from the platform
to the SDK. This should reduce bug reports for the SDK with respect to
unexpected response parsing. This PR also refactors the error
deserialization logic somewhat to make it more extensible in the future
for other potential error formats that are not currently handled.

## Breaking Changes
This PR renames MakeUnexpectedError() to MakeUnexpectedResponse() in the
`apierr` package. It also changes the return type to string. This makes
the message easier to incorporate into error responses that only depend
on the string representation of the error, as well as allows us to start
the message with a capital letter, as it is a complete sentence.

The error message for failed deserialization of valid responses has
changed slightly, from `unexpected error handling request` to `failed to
unmarshal response body`. The rest of the message is identical.

## Tests
Refactored unit tests to a table-driven test case, and added four new
cases: one for error details (not previously covered), one for the
regular happy path, one for unexpected responses, and one for the new
error message format.

- [ ] `make test` passing
- [ ] `make fmt` applied
- [ ] relevant integration tests applied

---------

Co-authored-by: Renaud Hartert <renaud.hartert@databricks.com>