[Fix] Fail fast when authenticating if host is not configured (#1033)
## Changes
The SDK currently hangs when making API requests if the Host is set to a
non-empty but invalid endpoint, such as `https://:443`. This PR adds a
check during authentication to ensure that the hostname is defined,
failing early if not. The underlying error is exported so other clients,
such as the CLI or Terraform Provider, can provide specific advice when
this happens.

## Tests
Added a unit test for this case.

- [ ] `make test` passing
- [ ] `make fmt` applied
- [ ] relevant integration tests applied