[Internal] Update Jobs GetRun API to support paginated responses for jobs and ForEach tasks (#1089)
## What changes are proposed in this pull request?

Introduces extension for jobs GetRun call that paginates tasks and
iterations arrays in the response. This change is necessary to prepare
for jobs API 2.2 release that serves paginated response. Pagination is
over once the `next_page_token` is absent from the response. The
pagination logic is not exposed to the customer.

## How is this tested?

Unit test and manual test. I wrote go code that called the new logic and
verified that the returned result contained full object when the API was
paginated.