[Internal] Decouple serving and oauth2 package (#1110)
## What changes are proposed in this pull request?

Package `oauth2` contains a `DataPlaneService` that is not intended for
users to use directly. This service is not auto-generated and is only
used in the `serving` package. To remove the dependency between the
`serving` package and the `oauth2` package, this PR moves the
`DataPlaneService` from the `oauth2` to the `serving` package.

## How is this tested?
Existing unit tests