[Internal] Move package credentials in config (#1115)
## What changes are proposed in this pull request?

This PR moves the `credentials` package in `config`, thus centralizing
the credentials logic in a single place.

The goal of this change is to prepare the current SDK-Beta to facilitate
migration to the new SDK that is currently being developed.

This PR introduces a breaking change by changing the import paths to
`CredentialsProvider`. However, we believe that the impact is fairly
contained as (i) `CredentialsProvider` is mostly internally, and (ii)
the breaking change can be addressed by updating the import path.

## How is this tested?

Through normal unit test and integration tests.