[Internal] Move all "core" packages in the "databricks" package.  (#1121)
## What changes are proposed in this pull request?

This PR moves all the "core" packages in the "databricks" directory
which is meant to become the core library.

This PR is also removing the mocked files while syncing with the latest
version of main.

Note that this PR also changes generated code. This is intended as the
code generator has to be updated to account for these changes.

## How is this tested?

Unit tests and integration tests are still passing.

---------

Co-authored-by: Parth Bansal <parth.bansal@databricks.com>