[Internal] Delete examples/mocking module (#1126)
## What changes are proposed in this pull request?
This PR deletes the `examples/mocking` directory as it is no longer
maintained. This mocker example is built on the assumption that every
`service` has `WithImpl()` that is used to set the stubs for every
`service`. But this is no longer the case. This mocker is not even able
to compile.

## How is this tested?
N/A