[Fix] Do not send query parameters when set to zero value (#1136)
## What changes are proposed in this pull request?

#1124 fixed the handling of APIs that use query parameters and a body
simultaneously. After that change, query parameters are always sent,
even if set to the zero value.

This PR addresses this, using the same logic for query parameters as for
fields in the body: if a field is set to its zero value, it will be
included in the query parameter if it is present in ForceSendFields.

## How is this tested?

This behavior is dependend upon in the Terraform provider. I'll be using
it in
https://github.com/databricks/terraform-provider-databricks/pull/4430
and verifying that there is no behavior change in the generated request.

When using this PR in Terraform, tests asserting that the path doesn't
include query parameters pass.