[Internal] Arrange services in a per package sdk approach (#1123)
## What changes are proposed in this pull request?

This PR moves all the service packages into separate Go modules and
isolates them from each other. It also removes the workspace and account
clients, which the per-service clients replace. It removes all the
examples as they use existing clients, so we need to revisit these in
the future.
Note: A few minor changes are also made, as we are removing the `go.mod`
from the root of the repository and moving things around.

## How is this tested?
Modified existing tests to use service clients instead of workspace and
account clients.