[Internal] Introduce new TokenSource interface that takes a `context.Context` (#1141)
## What changes are proposed in this pull request?

This PR adds a new `TokenSource` interface which is essentially the same
as `oauth2.TokenSource` but takes a `context.Context` as input.

The rationale behind this interface is that many of the SDK
`TokenSource` end-up making network calls under the hood. We want these
to be done passing the context of the calling method.

This interface is meant to ultimately entirely replace the use of
`oauth2.TokenSource` in the SDK.

## How is this tested?

Simple unit tests.