[Internal] Add support for asynchronous data plane token refreshes (#1142)
DO NOT MERGE: `api.go` and `impl.go` are here to help with review and
will be removed before merge. These are supposed to be added later on
via code generation.

## What changes are proposed in this pull request?

This PR is a step towards enabling asynchronous refreshes of data plane
tokens.

The PR introduces a new interface `dataplane.EndpointTokenSource` which
returns data plane tokens for a given endpoint and authorization detail.
The default implementation relies on `auth.NewCachedTokenSource` to wrap
the underlying TokenSources.

Note: async refreshes are disabled at the moment and will be enabled in
a follow-up PR.

**About Config.GetTokenSource():**

My first attempt to implement `GetTokenSource()` returned an error if
the `TokenSource` could not be built (e.g. because of config
initialization errors). This didn't work because:

1. service constructors do not return errors and thus had to panic;
2. service constructors are called systematically as soon as the
`WorkspaceClient` is created — sometimes with a config that is not
compatible with `GetTokenSource()`.

Returning a `TokenSource` provider instead of a `TokenSource` would have
solved the problem but did not feel right. The current solution (i.e.
return an ever-failing token source) is the cleanest solution I could
think off.

## How is this tested?

Complete unit test coverage of the new package.