[Feature] Enable async refreshes for OAuth tokens (#1143)
## What changes are proposed in this pull request?

This PR enables async refreshes for cached token sources. 

This PR also: 

- Fixes a race condition bug in the unit tests by adding a small delay
to the test refresh.
- Remove one test in AzureCLI that is not correct anymore due to how the
code is structured. Refactoring this to guarantee that the test pass is
not trivial and ultimately out of the scope of the new system — as the
cached token source is the one providing the guarantee.

## How is this tested?

Fix unit tests.