[Internal] Update Jobs `list_runs` function to support paginated responses (#1151)
## What changes are proposed in this pull request?

Introduces logic in the extension for jobs ListRuns call. The extended
logic accounts for the new response format of API 2.2. API 2.1 format
returns all tasks and job_cluster for each run in the runs list. API 2.2
format truncates tasks and job_cluster to 100 elements. The extended
ListRuns logic calls GetRun for each run in the list to populate the
full list of tasks and job_clusters.

I added logic that reads runs from the list and produces custom iterator
struct `expandedRunsIterator` that is supposed to mimic python
generators. The goal is to only read necessary elements from the API
endpoint and not more.

## How is this tested?

Unit tests and manual tests. Manual tests were done in two modes: using
API 2.2 and using API 2.1. So this code is compatible with both API
versions.

---------

Co-authored-by: Renaud Hartert <renaud.hartert@databricks.com>