[Internal] Initialize new `auth` module. (#1170)
## What changes are proposed in this pull request?

This PR creates the new `auth` module which is meant to become an
independent library to authenticate with Databricks.

The code is the same as what it was in `databricks/config/experimental`.
The only difference is that the cache now has its own package.

## How is this tested?

No change to the unit tests.

NO_CHANGELOG=true